+++ ios/Classes/SwiftDartQuickjsPlugin.swift
@@ -3,12 +3,8 @@ import UIKit
 
 public class SwiftDartQuickjsPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "dart_quickjs", binaryMessenger: registrar.messenger())
-    let instance = SwiftDartQuickjsPlugin()
-    registrar.addMethodCallDelegate(instance, channel: channel)
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    result("iOS " + UIDevice.current.systemVersion)
   }
 }