+++ lib/src/common.dart
@@ -4,6 +4,7 @@ import 'package:ffi/ffi.dart';
 import 'js_value.dart';
 import 'library.dart';
 import 'typedef.dart';
+import 'extensions.dart';
 
 class Common {
   Common._();
@@ -32,5 +33,30 @@ class Common {
     }
 
     return data;
+  }
+
+  // 执行构造函数
+  static Pointer<JSValue> callConstructor(
+    Pointer<JSContext> context,
+    Pointer<JSValue> constructor, [
+    List<JSObject>? args,
+  ]) {
+    if (library.isConstructor(context, constructor) == 0) {
+      throw 'constructor不是构造函数';
+    }
+
+    final pointer = library.callConstructor(
+      context,
+      constructor,
+      args?.length ?? 0,
+      args == null ? undefined : Common.formatArgs(args),
+    );
+
+    if (pointer.isException) {
+      final message = context.exception;
+      throw '函数调用执行异常:$message';
+    }
+
+    return pointer;
   }
 }