+++ macos/Classes/DartQuickjsPlugin.swift
@@ -3,17 +3,8 @@ import FlutterMacOS
 
 public class DartQuickjsPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "dart_quickjs", binaryMessenger: registrar.messenger)
-    let instance = DartQuickjsPlugin()
-    registrar.addMethodCallDelegate(instance, channel: channel)
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    switch call.method {
-    case "getPlatformVersion":
-      result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
-    default:
-      result(FlutterMethodNotImplemented)
-    }
   }
 }