+++ CHANGELOG.md
@@ -12,3 +12,9 @@
 * TODO: Addition of macOS platform, 
 * TODO: addition of bytecode compilation
 * TODO: Addition of bytecode execution.
+## 0.0.5
+* TODO: JSString adds `length` `indexOf` `lastIndexOf` `includes` `split` `concat`
+* TODO: JSArray adds `isArray` `join` `slice` `includes`
+* TODO: JSFunction adds the `off` method to remove subscriptions.
+* TODO: Update the JSFunction subscription logic to fix the issue where callback functions cannot be called when passed from Dart to JavaScript.
+* TODO: Add a new `JSRegExp` object.